% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\alias{ic.smfa}
\title{Extract information criteria of stochastic metafrontier models}
\usage{
\method{ic}{smfa}(object, IC = NULL, ...)
}
\arguments{
\item{object}{A stochastic metafrontier model returned
by \code{\link{smfa}}.}

\item{IC}{Character string. Information criterion measure. Three criteria
are available: \itemize{ \item \code{'AIC'} for Akaike information criterion
(default) \item \code{'BIC'} for Bayesian information criterion \item
\code{'HQIC'} for Hannan-Quinn information criterion }.}

\item{...}{Currently ignored.}
}
\value{
\code{\link{ic}} returns a data frame with the values of the
information criteria (AIC, BIC and HQIC) of the maximum likelihood coefficients.
If the \code{IC} argument is provided, it returns only the requested criterion
as a numeric value.
}
\description{
\code{\link{ic}} returns information criterion from stochastic
metafrontier models estimated with \code{\link{smfa}}.
}
\details{
The different information criteria are computed as follows: \itemize{ \item
AIC: \eqn{-2 \log{LL} + 2 * K} \item BIC: \eqn{-2 \log{LL} + \log{N} * K}
\item HQIC: \eqn{-2 \log{LL} + 2 \log{\left[\log{N}\right]} * K} } where
\eqn{LL} is the maximum likelihood value, \eqn{K} the number of parameters
estimated and \eqn{N} the number of observations.
}
\seealso{
\code{\link{smfa}}, for the stochastic metafrontier analysis model
fitting function using cross-sectional or pooled data.
}
\keyword{AIC}
\keyword{BIC}
\keyword{HQIC}
\keyword{methods}
