% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLik.smfa}
\title{Extract log-likelihood value of stochastic metafrontier models}
\usage{
\method{logLik}{smfa}(object, individual = FALSE, ...)
}
\arguments{
\item{object}{A stochastic metafrontier model returned
by \code{\link{smfa}}.}

\item{individual}{Logical. If \code{FALSE} (default), the sum of all
observations' log-likelihood values is returned. If \code{TRUE}, a vector of
each observation's log-likelihood value is returned.}

\item{...}{Currently ignored.}
}
\value{
\code{\link{logLik}} returns either an object of class
\code{'logLik'}, which is the log-likelihood value with the total number of
observations (\code{nobs}) and the number of free parameters (\code{df}) as
attributes, when \code{individual = FALSE}, or a list of elements, containing
the log-likelihood of each observation (\code{logLik}), the total number of
observations (\code{Nobs}) and the number of free parameters (\code{df}),
when \code{individual = TRUE}.
}
\description{
\code{\link{logLik}} extracts the log-likelihood value(s) from stochastic
metafrontier models estimated with \code{\link{smfa}}.
}
\seealso{
\code{\link{smfa}}, for the stochastic metafrontier analysis model
fitting function using cross-sectional or pooled data.
}
\keyword{likelihood}
\keyword{methods}
