% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs}
\alias{nobs}
\alias{nobs.smfa}
\title{Extract total number of observations used in frontier models}
\usage{
\method{nobs}{smfa}(object, ...)
}
\arguments{
\item{object}{a \code{smfa}
object for which the number of total observations is to be extracted.}

\item{\dots}{Currently ignored.}
}
\value{
A single number, normally an integer.
}
\description{
This function extracts the total number of 'observations' from a
fitted point frontier model.
}
\details{
\code{nobs} gives the number of observations actually
used by the estimation procedure.
}
\seealso{
\code{\link{smfa}}, for the stochastic metafrontier analysis model
fitting function using cross-sectional or pooled data
}
\keyword{attribute}
