% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\alias{residuals.smfa}
\title{Extract residuals of stochastic metafrontier models}
\usage{
\method{residuals}{smfa}(object, ...)
}
\arguments{
\item{object}{A stochastic metafrontier model returned
by \code{\link{smfa}}.}

\item{\dots}{Currently ignored.}
}
\value{
\code{\link{residuals}} returns a vector of residuals values.
}
\description{
This function returns the residuals' values from stochastic metafrontier models
estimated with \code{\link{smfa}}.
}
\note{
The residuals values are ordered in the same way as the corresponding
observations in the dataset used for the estimation.
}
\seealso{
\code{\link{smfa}}, for the stochastic metafrontier analysis model
fitting function using cross-sectional or pooled data.
}
\keyword{methods}
\keyword{residuals}
