% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.smfa}
\alias{print.summary.smfa}
\title{Summary of results for stochastic metafrontier models}
\usage{
\method{summary}{smfa}(object, ...)

\method{print}{summary.smfa}(x, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{object}{An object of class \code{'smfa'} returned by the
function \code{\link{smfa}}.}

\item{...}{Currently ignored.}

\item{x}{An object of class \code{'summary.smfa'}.}

\item{digits}{Numeric. Number of digits displayed in values.}
}
\value{
The \code{\link{summary}} method returns a list of class
\code{'summary.smfa'}
that contains the same elements as an object returned by \code{\link{smfa}}
with the following additional elements:

\item{AIC}{Akaike information criterion.}

\item{BIC}{Bayesian information criterion.}

\item{HQIC}{Hannan-Quinn information criterion.}

\item{metaRes}{Matrix of metafrontier estimates, their standard errors, z-values,
and asymptotic P-values.}

\item{effStats}{A list of efficiency statistics including group means and
class membership probabilities.}

\item{grpSummaries}{A list of summary objects for each group model.}
}
\description{
Create and print summary results for stochastic metafrontier models returned by
\code{\link{smfa}}.
}
\seealso{
\code{\link{smfa}}, for the stochastic metafrontier analysis model
fitting function for cross-sectional or pooled data.
}
\keyword{methods}
\keyword{summary}
