% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov}
\alias{vcov}
\alias{vcov.smfa}
\title{Compute variance-covariance matrix of stochastic metafrontier models}
\usage{
\method{vcov}{smfa}(object, ...)
}
\arguments{
\item{object}{A stochastic metafrontier model returned
by \code{\link{smfa}}.}

\item{...}{Currently ignored}
}
\value{
The variance-covariance matrix of the maximum likelihood
coefficients is returned.
}
\description{
\code{\link{vcov}} computes the variance-covariance matrix of the maximum
likelihood (ML) coefficients from stochastic metafrontier models estimated with
\code{\link{smfa}}.
}
\details{
The variance-covariance matrix is obtained by the inversion of the
negative Hessian matrix. Depending on the distribution and the
\code{'hessianType'} option, the analytical/numeric Hessian or the bhhh
Hessian is evaluated.
}
\seealso{
\code{\link{smfa}}, for the stochastic metafrontier analysis model
fitting function using cross-sectional or pooled data.
}
\keyword{methods}
\keyword{vcov}
